Add LocationServiceUtils and move all location related functionality to a common place
