+++ app/src/main/java/com/deltasoft/pharmatracker/MainActivity.kt
@@ -43,7 +43,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.deltasoft.pharmatracker.navigation.Screen
-import com.deltasoft.pharmatracker.screens.home.location.LocationPingService
+import com.deltasoft.pharmatracker.screens.home.location.LocationServiceUtils
 import com.deltasoft.pharmatracker.screens.home.trips.ScheduledTripsState
 import com.deltasoft.pharmatracker.utils.AppUtils
 import com.deltasoft.pharmatracker.utils.sharedpreferences.PrefsKey
@@ -171,15 +171,12 @@ class MainActivity : ComponentActivity() {
                                     scheduledTripsResponse?.trips?.any { it?.status.equals("STARTED") }?:false
                                 Log.d("VMListener", "anyTripIsCurrentlyActive $anyTripIsCurrentlyActive")
                                 if (anyTripIsCurrentlyActive){
-//                                    if (!LocationPingService.isServiceRunning) {
-//                                        AppUtils.restartForegroundService(applicationContext)
-//                                    }
                                     viewModel.clearScheduledTripsState()
-                                    if(!LocationPingService.isServiceRunning) {
-                                        AppUtils.restartForegroundService(applicationContext)
+                                    if(LocationServiceUtils.isLocationServiceNotRunning()) {
+                                        LocationServiceUtils.restartForegroundService(applicationContext)
                                     }
                                 }else{
-                                    AppUtils.stopService(applicationContext)
+                                    LocationServiceUtils.stopService(applicationContext)
                                 }
                             }
                             is ScheduledTripsState.Error -> {
@@ -291,7 +288,7 @@ class MainActivity : ComponentActivity() {
                 )
             }
         }
-        if (isFineLocationPermissionGranted(this@MainActivity) && LocationPingService.isServiceRunning) {
+        if (isFineLocationPermissionGranted(this@MainActivity) && LocationServiceUtils.isLocationServiceRunning()) {
             viewModel?.checkAndSendLocationToServer(TAG, restartService = false)
         }
     }
